pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "path-term-kit"
version = "0.1.0"
description = "Reusable pathology subspecialty terminology consensus SOP toolkit."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Pathology Terminology Consensus Kit Maintainers" }
]
dependencies = [
  "openpyxl>=3.1.2",
  "PyYAML>=6.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "jsonschema>=4.22"
]

[project.scripts]
path-term-kit = "path_term_kit.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
